Change transactions number to integer, Moved validators to DTOs

+++ analytics-service/src/main/java/cz/muni/fi/obs/api/DailySummary.java
@@ -4,8 +4,8 @@ import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record DailySummary(LocalDate date,
-                           BigDecimal totalWithdrawalTransactions,
-                           BigDecimal totalDepositTransactions,
+                           Integer totalWithdrawalTransactions,
+                           Integer totalDepositTransactions,
                            BigDecimal totalWithdrawn,
                            BigDecimal totalDeposited,
                            BigDecimal averageWithdrawn,
