+++ analytics-service/src/main/java/cz/muni/fi/obs/api/MonthlySummary.java
@@ -3,8 +3,8 @@ package cz.muni.fi.obs.api;
 import java.math.BigDecimal;
 
 public record MonthlySummary(String month,
-                             BigDecimal totalWithdrawalTransactions,
-                             BigDecimal totalDepositTransactions,
+                             Integer totalWithdrawalTransactions,
+                             Integer totalDepositTransactions,
                              BigDecimal totalWithdrawn,
                              BigDecimal totalDeposited,
                              BigDecimal averageWithdrawn,
